requirements.txt
torch
torchvision
pandas
grad-cam
thop
pthflops
pyyaml